Add ListBuilder.
This commit adds a builder for more efficiently creating `List`
objects. It works by mutating the list in-place until the current
list is requested. After that request, the next mutation will occur
on a copy but then continue to mutate in place. This allows lists
to be returned and used while continuing to use the builder.